+++ packages/native/src/components/RuleNative.tsx
@@ -2,9 +2,11 @@ import * as React from 'react';
 import { useMemo } from 'react';
 import { StyleSheet, View } from 'react-native';
 import type { FullField } from 'react-querybuilder';
-import { RuleComponents, TestID, useRule } from 'react-querybuilder';
+import { RuleComponents, RuleComponentsWithSubQuery, TestID, useRule } from 'react-querybuilder';
 import { defaultNativeStyles } from '../styles';
 import type { RuleNativeProps, SchemaNative } from '../types';
+
+const GroupComponentsWrapper = (props: React.PropsWithChildren<unknown>) => <View {...props} />;
 
 /**
  * @group Components
@@ -20,7 +22,11 @@ export const RuleNative = (props: RuleNativeProps): React.JSX.Element => {
 
   return (
     <View style={styles.rule} testID={TestID.rule}>
-      <RuleComponents {...r} />
+      {r.matchModes.length > 0 ? (
+        <RuleComponentsWithSubQuery {...r} groupComponentsWrapper={GroupComponentsWrapper} />
+      ) : (
+        <RuleComponents {...r} />
+      )}
     </View>
   );
 };