+++ packages/react-querybuilder/src/components/QueryBuilder.useQueryBuilderSetup.ts
@@ -1,5 +1,6 @@
 import { useCallback, useMemo, useState } from 'react';
 import { defaultCombinators, defaultOperators } from '../defaults';
+import { useFields } from '../hooks';
 import type { UseMergedContextReturn } from '../hooks/useMergedContext';
 import { useMergedContext } from '../hooks/useMergedContext';
 import type {
@@ -11,9 +12,9 @@ import type {
   FullOption,
   FullOptionList,
   FullOptionMap,
-  FullOptionRecord,
   GetOptionIdentifierType,
   GetRuleTypeFromGroupWithFieldAndOperator,
+  MatchModeOptions,
   Option,
   OptionGroup,
   QueryBuilderProps,
@@ -26,15 +27,12 @@ import {
   filterFieldsByComparator,
   generateID,
   getFirstOption,
+  getMatchModesUtil,
+  getOption,
   getValueSourcesUtil,
   isFlexibleOptionGroupArray,
   joinWith,
-  objectKeys,
-  toFullOption,
   toFullOptionList,
-  toFullOptionMap,
-  uniqByIdentifier,
-  uniqOptGroups,
   uniqOptList,
 } from '../utils';
 
@@ -65,12 +63,12 @@ export type UseQueryBuilderSetup<
   rqbContext: UseMergedContextReturn<F, GetOptionIdentifierType<O>, true>;
   fields: FullOptionList<F>;
   fieldMap: FullOptionMap<
-    FullField<string, string, string, Option<string>, Option<string>>,
+    FullField<string, string, string, FullOption, FullOption>,
     GetOptionIdentifierType<F>
   >;
   combinators:
-    | WithUnknownIndex<BaseOption<string> & FullOption<string>>[]
-    | OptionGroup<WithUnknownIndex<BaseOption<string> & FullOption<string>>>[];
+    | WithUnknownIndex<BaseOption<string> & FullOption>[]
+    | OptionGroup<WithUnknownIndex<BaseOption<string> & FullOption>>[];
   getRuleDefaultValue: <RT extends RuleType = GetRuleTypeFromGroupWithFieldAndOperator<RG, F, O>>(
     r: RT
   ) => any; // eslint-disable-line @typescript-eslint/no-explicit-any
@@ -82,6 +80,14 @@ export type UseQueryBuilderSetup<
   getValueEditorTypeMain: QueryBuilderProps<RG, F, O, C>['getValueEditorType'];
   getValueSourcesMain: QueryBuilderProps<RG, F, O, C>['getValueSources'];
 }> & {
+    getSubQueryBuilderPropsMain: (
+      field: GetOptionIdentifierType<F>,
+      misc: { fieldData: F }
+    ) => Record<string, unknown>;
+    getMatchModesMain: (
+      field: GetOptionIdentifierType<F>,
+      misc?: { fieldData: F }
+    ) => MatchModeOptions;
     getOperatorsMain: (
       ...p: Parameters<NonNullable<QueryBuilderProps<RG, F, O, C>['getOperators']>>
     ) => FullOptionList<O>;
@@ -111,7 +117,7 @@ export const useQueryBuilderSetup = <
   const [qbId] = useState(generateID);
 
   const {
-    fields: fieldsPropOriginal,
+    fields: fieldsProp,
     baseField,
     operators: operatorsProp,
     baseOperator,
@@ -124,7 +130,9 @@ export const useQueryBuilderSetup = <
     getDefaultField,
     getDefaultOperator,
     getDefaultValue,
+    getMatchModes,
     getOperators,
+    getSubQueryBuilderProps,
     getValueEditorType,
     getValueSources,
     getInputType,
@@ -158,84 +166,12 @@ export const useQueryBuilderSetup = <
   const { translations } = rqbContext;
 
   // #region Set up `fields`
-  const defaultField = useMemo(
-    () =>
-      ({
-        id: translations.fields.placeholderName,
-        name: translations.fields.placeholderName,
-        value: translations.fields.placeholderName,
-        label: translations.fields.placeholderLabel,
-      }) as FullField,
-    [translations.fields.placeholderLabel, translations.fields.placeholderName]
-  );
-  const fieldsProp = useMemo(
-    () => fieldsPropOriginal ?? ([defaultField] as FlexibleOptionList<F>),
-    [defaultField, fieldsPropOriginal]
-  );
-
-  const fields = useMemo((): F[] | OptionGroup<F>[] => {
-    const flds = (
-      Array.isArray(fieldsProp)
-        ? toFullOptionList(fieldsProp, baseField)
-        : (objectKeys(toFullOptionMap(fieldsProp, baseField)) as unknown as FieldName[])
-            .map(fld => ({ ...fieldsProp[fld], name: fld, value: fld }))
-            .sort((a, b) => a.label.localeCompare(b.label))
-    ) as FullOptionList<F>;
-    if (isFlexibleOptionGroupArray(flds)) {
-      return autoSelectField
-        ? (uniqOptGroups(flds) as FullOptionList<F>)
-        : (uniqOptGroups([
-            {
-              label: translations.fields.placeholderGroupLabel,
-              options: [defaultField],
-            },
-            ...flds,
-          ]) as FullOptionList<F>);
-    } else {
-      return autoSelectField
-        ? (uniqByIdentifier(flds as F[]) as FullOptionList<F>)
-        : (uniqByIdentifier([defaultField, ...(flds as F[])]) as FullOptionList<F>);
-    }
-  }, [
-    autoSelectField,
+  const { fields, fieldMap } = useFields({
+    fields: fieldsProp,
     baseField,
-    defaultField,
-    fieldsProp,
-    translations.fields.placeholderGroupLabel,
-  ]);
-
-  const fieldMap = useMemo(() => {
-    if (!Array.isArray(fieldsProp)) {
-      const fp = toFullOptionMap(fieldsProp, baseField) as FullOptionMap<FullField, FieldName>;
-      return autoSelectField ? fp : { ...fp, [translations.fields.placeholderName]: defaultField };
-    }
-    const fm: Partial<FullOptionRecord<FullField>> = {};
-    if (isFlexibleOptionGroupArray(fields)) {
-      for (const f of fields) {
-        for (const opt of f.options) {
-          fm[(opt.value ?? /* istanbul ignore next */ opt.name) as FieldName] = toFullOption(
-            opt,
-            baseField
-          ) as FullField;
-        }
-      }
-    } else {
-      for (const f of fields) {
-        fm[(f.value ?? /* istanbul ignore next */ f.name) as FieldName] = toFullOption(
-          f,
-          baseField
-        ) as FullField;
-      }
-    }
-    return fm;
-  }, [
     autoSelectField,
-    baseField,
-    defaultField,
-    fields,
-    fieldsProp,
-    translations.fields.placeholderName,
-  ]);
+    translations,
+  });
   // #endregion
 
   const combinators = useMemo(
@@ -331,6 +267,19 @@ export const useQueryBuilderSetup = <
     (field: FieldName, operator: OperatorName) =>
       getValueSourcesUtil<F, OperatorName>(fieldMap[field] as F, operator, getValueSources),
     [fieldMap, getValueSources]
+  );
+
+  const getMatchModesMain = useCallback(
+    (field: FieldName, _misc?: { fieldData: F }) =>
+      getMatchModesUtil<F>(fieldMap[field] as F, getMatchModes),
+    [fieldMap, getMatchModes]
+  );
+
+  const getSubQueryBuilderPropsMain = useCallback(
+    (field: FieldName, misc: { fieldData: F }): Record<string, unknown> =>
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      getSubQueryBuilderProps?.(field, misc) ?? ({} as any),
+    [getSubQueryBuilderProps]
   );
 
   const defaultValueOption = useMemo(
@@ -452,12 +401,17 @@ export const useQueryBuilderSetup = <
 
     const valueSource = getValueSourcesMain(field, operator)[0] ?? 'value';
 
+    const matchMode = getFirstOption(
+      getMatchModesMain(field, { fieldData: getOption(flds, field) as F })
+    );
+
     const newRule = {
       id: idGenerator(),
       field,
       operator,
       valueSource,
       value: '',
+      ...(matchMode ? { match: { mode: matchMode, threshold: 1 } } : null),
     } as unknown as R;
 
     const value = getRuleDefaultValue(newRule);
@@ -466,6 +420,7 @@ export const useQueryBuilderSetup = <
   }, [
     fields,
     getDefaultField,
+    getMatchModesMain,
     getRuleDefaultOperator,
     getRuleDefaultValue,
     getValueSourcesMain,
@@ -501,8 +456,10 @@ export const useQueryBuilderSetup = <
     fields: fields as FullOptionList<F>,
     fieldMap,
     combinators,
+    getMatchModesMain,
     getOperatorsMain,
     getRuleDefaultOperator,
+    getSubQueryBuilderPropsMain,
     getValueEditorTypeMain,
     getValueSourcesMain,
     getValuesMain,