+++ packages/react-querybuilder/src/hooks/index.ts
@@ -1,5 +1,6 @@
 export * from './useControlledOrUncontrolled';
 export * from './useDeprecatedProps';
+export * from './useFields';
 export * from './useMergedContext';
 export * from './usePreferProp';
 export * from './usePrevious';