+++ packages/react-querybuilder/src/utils/mergeClassnames.ts
@@ -43,6 +43,8 @@ export const mergeClassnames = (...args: MergeClassnamesParams): Classnames => {
     dndGroup: joinClassnamesByName('dndGroup', args),
     disabled: joinClassnamesByName('disabled', args),
     valueListItem: joinClassnamesByName('valueListItem', args),
+    matchMode: joinClassnamesByName('matchMode', args),
+    matchThreshold: joinClassnamesByName('matchThreshold', args),
     branches: joinClassnamesByName('branches', args),
   };
 };